Final assignment, caesar and matrix crypt. Added conversion to and from binary file once encryption/decryption is done.
