+++ sabinpruna/concurrency/wordCountSequential.go
